feat(server/pypika): Add support for the duration step
closes TCTCC-6485
